removed recursion caused by taxItems saveing the cart.  cleaned up the checkout controller and added a helper function to the customers model.  increased customer counts after order is accepted, fixes dioscouri/amritasingh-f3#184

+++ src/Shop/Models/Customers.php
@@ -109,9 +109,9 @@ class Customers extends \Users\Models\Users
         	return (int) $this->{'shop.orders_count'};
         }
         
-        $this->{'shop.orders_count'} = (new \Shop\Models\Orders)->setState('filter.user', $this->id)->setState('filter.financial_status', \Shop\Constants\OrderFinancialStatus::paid)->getCount();
+        $this->{'shop.orders_count'} = (int) (new \Shop\Models\Orders)->setState('filter.user', $this->id)->setState('filter.financial_status', \Shop\Constants\OrderFinancialStatus::paid)->getCount();
         
-        return (int) $this->{'shop.orders_count'};
+        return $this->{'shop.orders_count'};
     }
     
     /**
